Remove unused kwset.[ch]
Remove the unused kwset matching engine. It has not been used by grep
itself since 48de2a768cf (grep: remove the kwset optimization,
2019-07-01), and in a previous commit we removed its last remaining
use in diffcore-pickaxe.c.

This matching engine is much faster than the C library's regex
matching for fixed strings, but slower than PCREv2, which we can make
use of in the common case.

It's also an increasingly bitrotting version of the GPLv2 code, whose
upstream has moved onto the incompatible GPLv3, so getting rid of the
need to maintain a permanent fork of it is a good thing.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>