Makefile/coccicheck: speed up and fix bug with duplicate hunks
Change the coccicheck target to run on all of our *.c and *.h files
with --include-headers-for-types, instead of trusting it to find *.h
files and other includes to modify from its recursive walking of
includes as it has been doing with only --all-includes.

TODO

This speeds it up significantly and reduces its memory usage, since it
doesn't need to parse N includes for every file it visits.

See [1] for a discussion thread about this commit with some timings
for details, but briefly: This change speeds it up by ~2x and makes it
use much less memory. Or a reduction of a max of around ~2GB
per-process (under the old SPATCH_BATCH_SIZE=0) to around ~200MB.

Looking at the history of the coccicheck target I think we've always
been running it in the wrong mode for what we wanted to achieve:

 - When it was added in 63f0a758a06 (add coccicheck make target,
   2016-09-15) it explicitly processed only the %.c files.

 - Then in a9a884aea57 (coccicheck: use --all-includes by default,
   2016-09-30) it started processing the %.h files by looking around for
   its own includes.

Let's instead just point it to both our *.c and *.h files, then
there's no need to have it recursively look around for included files
to change.

Setting --no-includes would not work if we expected to set
COCCI_SOURCES to a subset of our source files, but that's not what
we're doing here. If anyone manually tweaks COCCI_SOURCES they'll now
need to tweak SPATCH_FLAGS too. The speed and correctness we gain is
worth that small trade-off.

Using --no-includes also fixes a subtle bug introduced in
960154b9c17 (coccicheck: optionally batch spatch invocations,
2019-05-06) with duplicate hunks being written to the
generated *.patch files.

This is because that change altered a file-at-a-time for-loop to an
invocation of "xargs -n X". This would not matter for most other
programs, but it matters for spatch.

This is because each spatch invocation will maintain shared lock files
in /tmp, check if files being parsed were changed etc. I haven't dug
into why exactly, but it's easy to reproduce the issue[2]. The issue
goes away entirely if we just use --no-includes, which as noted above
would have made sense even without that issue.

1. https://lore.kernel.org/git/20210302205103.12230-1-avarab@gmail.com/
2. A session showing racy spatch under xargs -n X:

    $ cat test.cocci
    @@
    expression E1;
    @@
    - strbuf_avail(E1)
    + strbuf_has(E1)
    $ for i in 1 2 4 16 64 128 512
    do
        echo with xargs -n $i: &&
	echo *.c | xargs -n $i spatch --sp-file \
            test.cocci --all-includes --patch . 2>/dev/null | \
	grep -F +++ | sort | uniq -c
    done
    with xargs -n 1:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
    with xargs -n 2:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
    with xargs -n 4:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
    with xargs -n 16:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
          2 +++ b/strbuf.h
    with xargs -n 64:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
          2 +++ b/strbuf.h
    with xargs -n 128:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
          2 +++ b/strbuf.h
    with xargs -n 512:
          1 +++ b/convert.c
          1 +++ b/strbuf.c
          1 +++ b/strbuf.h

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>

Makefile/coccicheck: use --include-headers-for-types

Change the recently added --no-includes out for --all-includes
--include-headers-for-types.

When I moved from --all-includes to --no-includes I missed that rules
that use "type T" such as the one added dbc540c7a58 (add QSORT,
2016-09-29) won't work as intended. See [1] for a report of that.

So let's move back to --all-includes to potentially scour our includes
for definitions, but supply --include-headers-for-types so we won't
racily emit duplicate hunks. See [2] for more discussion about that
issue.

This undoes most of the optimization we gained with --no-includes,
using a very basic benchmark script [3] I get these results:

 * ~1m50s: --all-includes (before my recent patches)
   38 files changed, 74 insertions(+), 78 deletions(-)
 * ~55s:   --no-includes
   27 files changed, 55 insertions(+), 57 deletions(-)
 * ~1m35s: --all-includes --include-headers-for-types
   38 files changed, 74 insertions(+), 78 deletions(-)
 * ~1m20s: --local-includes --include-headers-for-types
   36 files changed, 70 insertions(+), 72 deletions(-)
 * ~1m20s: --local-includes --include-headers-for-types --no-loops
   36 files changed, 70 insertions(+), 72 deletions(-)

And only the "--all-includes --include-headers-for-types" gets the
same results as the previous "--all-includes", even with
"--local-includes" we miss out on some definitions.

1. https://lore.kernel.org/git/3aac381e-2ce9-e35e-498c-9c26df235aed@web.de/
2. https://lore.kernel.org/git/YENdUMLTM+cerfqJ@coredump.intra.peff.net/
3.
   for flags in \
       '--all-includes' \
       '--no-includes' \
       '--all-includes --include-headers-for-types' \
       '--local-includes --include-headers-for-types' \
       '--local-includes --include-headers-for-types --no-loops'
   do
   	git reset --hard &&
           git clean -dxfq contrib/
   	time make -j8 coccicheck SPATCH_FLAGS="$flags --patch ." SPATCH_XARGS="xargs -P 8 -n 16" 2>&1 | grep -v SPATCH
   	cat contrib/coccinelle/*.patch | git apply
   	git --no-pager diff --shortstat
   	git --no-pager diff | git patch-id
   done

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>