tree-walk.h API users: rename "struct name_entry"'s "mode" to "raw_mode"
Now that most of the users of the "mode" variable have been moved to
use "object_type" instead let's rename it to "raw_mode" in preparation
for a revert of 7146e66f086 (tree-walk: finally switch over tree
descriptors to contain a pre-parsed entry, 2014-02-06).

This will allow API users who care about the actual mode bits in tree
objects to get access to them, such as fsck, the merge algorithm etc.

But most users will not want to have such potentially un-sanitized, so
let's indicate that by giving the variable a more scary name.

I am not renaming the variables being assigned to, i.e. it's now going
to be "int mode = entry.raw_mode", not "int raw_mode = [...]". This is
because we're going to be getting a sanitized "mode" via
"canon_mode()" in many of these functions soon, so renaming the local
variable back and forth will result in needless churn.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>