test-lib.sh: capitalize the "skip" TAP directive
Correct a minor error in the TAP output. When I added the TAP output
in 5099b99d25f (test-lib: Adjust output to be valid TAP format,
2010-06-24) I introduced a mixture of "# SKIP" and "# skip".

The TAP specification is rather strict on the matter, and technically
these shouldn't be counted[1], but in practice the prove(1) utility
accepts these case-insensitively[2].

However, prove will helpfully normalize these with "--normalize", and
we're not guaranteed that this'll work with all TAP consumers, so
let's tighten this up.

1. https://testanything.org/tap-specification.html#directives
2. https://metacpan.org/source/TAP::Parser::Grammar

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>