test-tool: add a "tee"-alike "tee-tap" tool
Add a tool similar to "tee" to be used to re-implement the
"--verbose-log" option. Currently that relies on juggling stdout and
stderr, I'm going to switch it to using this tool that understands the
TAP format, and can selectively print things from a TAP stream to
stdout, while writing the full verbose test log to a file.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>