test-lib tests: add a test for test-lib.sh output variety
Add a new test which compliments the tests of test-lib.sh
functionality in t0000-basic.sh, it's in a separate file to make
output changes easier to review and consider, whereas the tests in
t0000-basic.sh are more focused on stess testing the functionality of
the test framework.

The test runs successfully, but can be made to fail with real failing
tests by supplying TEST_LIB_OUTPUT_DEMO=true. It can thus be used to
check the output in a way that e.g. emits color to the user's
terminal, without having to worry about this being a test-in-a-test
using only lib-subtest.sh, with all the associated output that setup
would add.

The passing output looks like this:

    $ ./t0150-test-lib-output-demo.sh
    Set TEST_LIB_OUTPUT_DEMO=true to see a test-lib.sh output demo
    ok 1 - successful one-line test
    ok 2 - successful two-line test
    ok 3 - unexpectedly passing TODO test
    not ok 4 - failing TODO test # TODO known breakage
    ok 5 # skip we will be skipping this test (GIT_SKIP_TESTS)
    not ok 6 - an one-line fail (set TEST_LIB_OUTPUT_DEMO=true) # TODO known breakage
    not ok 7 - a multi-line failure (set TEST_LIB_OUTPUT_DEMO=true) # TODO known breakage
    ok 8 - setup t0150-fake.sh
    ok 9 - run t0150-fake.sh
    ok 10 - run t0150-fake.sh --verbose
    # still have 3 known breakage(s)
    # passed all remaining 7 test(s)
    1..10
    $ echo $?
    0

But the test can also be made to fail, which demonstrates the output
emitted on failure:

    $ TEST_LIB_OUTPUT_DEMO=true ./t0150-test-lib-output-demo.sh
    ok 1 - successful one-line test
    ok 2 - successful two-line test
    ok 3 - unexpectedly passing TODO test # TODO known breakage vanished
    not ok 4 - failing TODO test # TODO known breakage
    ok 5 # skip we will be skipping this test (GIT_SKIP_TESTS)
    not ok 6 - an one-line fail (set TEST_LIB_OUTPUT_DEMO=true)
    #       false
    not ok 7 - a multi-line failure (set TEST_LIB_OUTPUT_DEMO=true)
    #
    #               test_when_finished "rm out" &&
    #               # will not be empty!
    #               echo >out &&
    #               test_must_be_empty out
    #
    ok 8 - setup t0150-fake.sh
    ok 9 - run t0150-fake.sh
    ok 10 - run t0150-fake.sh --verbose
    # 1 known breakage(s) vanished; please update test(s)
    # still have 1 known breakage(s)
    # failed 2 among remaining 8 test(s)
    1..10
    $ echo $?
    1

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>