test-lib: add --color to force color output
Add a positive --color option to compliment the existing --no-color
option added in 55db1df0c86 (Add some fancy colors in the test library
when terminal supports it., 2007-10-24).

This will be used to test the colors the test-lib.sh itself outputs,
and is also useful if for whatever reason we'd like to pipe its output
somewhere and preserve the coloring.

In order to support that we need to move away from using tput to
printing the color codes directly. The tput command will print nothing
if TERM=dumb (or equivalent). Since there was no --no-color option
before this wasn't an issue, now it is.

This doesn't introduce over-coloring when not under --color, since
we'll use tput just above the second hunk being modified here to set
the $color default.

It is important though that these assignments happen outside the
say_color() function, see d5c1b7c286b (test-lib.sh: fix color support
when tput needs ~/.terminfo, 2015-06-17), but that isn't being changed
here.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>