test-lib tests: pass test output through test_decode_color
Pass the output through test_decode_color in preparation for a
subsequent change to test color output from test-lib.sh. I've split
this up into another change to show that's it's functionally
equivalent, and because of the relative verbosity of the X -> X.raw
change.

I could have added the decoding to _run_sub_test_lib_test_common(),
but that would only appease these tests which peek behind the
curtains. The relevant parts of them will mostly be rewritten soon as
I'm changing this part of the test output, so let's keep this
intermediate step.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>