test-lib: reformat argument list in test_create_repo()
Reformat an argument list changed in 675704c74dd (init: provide useful
advice about init.defaultBranch, 2020-12-11) to have the "-c" on the
same line as the argument it sets.

If and when that line is removed it's going to be less readable to
have to see two lines changed v.s. one.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>