test-lib: make "git init" in test_create_repo "--quiet"
Change the "git init" invocation in test_create_repo() to be quiet, it
will no longer emit a notice about an empty repository being created
as the first line in the --verbose output of most tests (except those
that set TEST_NO_CREATE_REPO).

It is useful to have this message in one form or another, but in
subsequent commits I'll be explicitly testing the full output of the
test suite under --verbose.

As hinted at in the test I'm altering here in t0000-basic.sh the
verbose output of test_create_repo() cannot be fed to test_cmp,
because under --root=<path> the full filesystem path to the repository
will be emitted.

I'll be bringing back a better version of this debugging output, but
only after I've set up some output testing, utility functions for
emitting better output etc.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>