test-lib: make "git init" in test_create_repo "--quiet"
Change the "git init" invocation in test_create_repo() to be quiet, it
will no longer emit a notice about an empty repository being created
as the first line in the --verbose output of most tests (except those
that set TEST_NO_CREATE_REPO).

It is apparent that a repository is being created from anyone with
familiarity with the basics of the test framework. To the extent that
this emission to stderr is useful to anyone (e.g. a reminder of what
--root was picked?) it'll be brought back in subsequent commits in the
form of one of the notices a configurable TAP verbosity level emits.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>