test-lib output: color only the TODO part yellow
In a preceding change I changed the "ok" output for "not ok ... #
TODO" tests to be green, as part of making all passing tests
consistently green.

This undid a fix to the output done in e8e5195573b (tests: paint known
breakages in yellow, 2012-12-16).

Let's follow-up and have our cake and eat it too. Now we consistently
color the expected "ok/not ok" portion, but will color the "TODO" part
yellow.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>