test-lib: add say_color_tap_comment() helper to emit TAP comments
Add a say_color_tap_comment() helper and use it for emitting the TAP
comments we print before the test count at the end.

As alluded to in the comment being added here this is the start of a
more general facility for emitting N-level TAP comments. Eventually
I'll be changing the --verbose-log mode to support printing comments
up to level N.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>