test-lib tests: add tests for skip_all output
Add tests for the output emitted under skip_all and when
GIT_SKIP_TESTS matches the test name itself.

The first use of skip_all was in my fadb5156e48 (tests: Skip tests in
a way that makes sense under TAP, 2010-06-24), after I'd added the
facility in 04ece59399b (GIT_SKIP_TESTS: allow users to omit tests
that are known to break, 2006-12-28).

But the GIT_SKIP_TESTS method of doing this dates back to
5099b99d25f (test-lib: Adjust output to be valid TAP format,
2010-06-24).

As seen in these tests the output is subtly different, let's test for
it for now. A later commit will amend the output to be consistent.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>