test-lib: change output for consistency and parsing
Change the output emitted from test-lib.sh to make it more
consistent (e.g. in its use of coloring), and so that the source of
tests is reported after the "ok/not ok" line instead of before it.

As noted in an earlier commit the end goal here is to have N-level #
comments representing different levels of TAP verbosity. See the
updated documentation in test-lib.sh for what the verbosity levels
mean.

This does mean that e.g. if a test sleeps for 10 seconds instead of
showing the equivalent of:

    about to run: sleep 10
    [sleep for 10]
    ok 1 - slept for 10

We'll instead show

    [sleep for 10]
    ok 1 - slept for 10
    ### sleep 10

I think that's a good trade-off in the common case. We can still show
what's happening with "-x", but having the source of the failed test
after the "ok/not ok" line makes things easier to parse, and opens the
door to e.g. annotating the specific failed line in a &&-chain without
requiring extra output.

I'm also removing the auto-indentation that's been done ever since
bf0dd8a8de9 ([PATCH] Test framework: prettyprint the failed command.,
2005-07-22). I think in practice it was counter-productive and
resulted in uglier output in the common case.

The reason the diff to t0000-basic.sh here looks uglier after this
change is because those tests don't represent the common case. Due to
how shell "<<-EOF" here-docs work the second TAB in those tests is
stripped away, and thus e.g. "false" will be at the beginning of the
line.

But for almost all of the rest of our tests we can rely on the test
itself coming with indentation, see the change to
t0150-test-lib-output-demo.sh being made here for a better example. As
can be seen in that case we won't double-indent the source anymore,
nor add lines consisting only of tabbed whitespace.

I've also made the output more color-consistent, e.g. the body of
skipped test scripts is now colored the same as the TAP directive for
it, so are the summary lines at the end.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>