test-lib tests: add a TODO test for a --verbose-only bug
Test for a bug that's been present in the --verbose-only mode ever
since ff09af3fb8f (test-lib: verbose mode for only tests matching a
pattern, 2013-06-23).

Under --verbose we emit an extra newline after running tests to
prevent test output interfering with the TAP output. See
57e1538ac9b (test-lib: output a newline before "ok" under a TAP
harness, 2010-06-24).

But the --verbose-only mode did not account for this, as demonstrated
by this test.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>