test-lib: make --verbose-only match test names, not file names
Change the --verbose-only option added in ff09af3fb8f (test-lib:
verbose mode for only tests matching a pattern, 2013-06-23) so that
its argument matches test names, and doesn't glob whatever files
happen to be in the current directory.

It was only an accident running it as e.g. --verbose-only=1* would
work most of the time, because the common case was that there would be
no files in the trash directory matching that pattern.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>