test-lib: make the --valgrind-only option match test names, not file names
Fix a bug in the --valgrind-only option, whose setup code was
copy/pasted from the --verbose-only option. See the preceding bug fix
for --verbose-only for a demonstration of the bug.

I'm not adding a test here because I can't be bothered to figure out
how to properly guard the valgrind setup in t0000-basic.sh for this
minor bug, there are no existing valgrind tests in the file.

This bug dates back to 5dfc368f5ec (test-lib: valgrind for only tests
matching a pattern, 2013-06-23).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>