githooks doc: add a reference to uploadpack.packObjectsHook
Add a reference to the uploadpack.packObjectsHook hook to githooks(5),
for those readers who might be looking for a hook for pack-objects.

I was sure there was such a hook, but finding the docs took me a while
because I started out searching for "pack-objects" in the githooks(5)
documentation. Let's aid future readers with the same confusion.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>