git-send-email doc: discuss all config options in git-config(1)
It's bothered me for a while that git-send-email(1) is a special
snowflake in that its configuration options weren't fully discussed in
either its own documentation, or git-config(1).

This change rectifies that. We now have the canonical documentation in
git-config(1), and include it in git-send-email(1).

This is mostly a move-only change, the exception is a minor rewording
of changing wording like "see above" to "see linkgit:git-config[1]",
as well as a clarification about the big section of command-line
option tweaking config being discussed in git-send-email(1)'s main
docs.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>