branch docs: include the branch.* section from config
Include the branch.* section from git-config(1) in the git-branch
docs, similarly to what I did in b6a8d09f6d8 (gc docs: include the
"gc.*" section from "config" in "gc", 2019-04-07).

Unlike the just-introduced template for saying the entire section is
included, in this case we have some discussion unique to the
documentation earlier in it. Let's add a standard text for that into
includes/* for later use in other similar scenarios.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>