fsck: don't hard die on invalid object types
Change builtin/fsck.c to pass down a
OBJECT_INFO_ALLOW_UNKNOWN_TYPE. This changes this very ungraceful
error:

    $ git hash-object --stdin -w -t garbage --literally </dev/null
    <OID>
    $ git fsck
    fatal: invalid object type
    $

Into:

    $ git fsck
    error: hash mismatch for <OID_PATH> (expected <OID>)
    error: <OID>: object corrupt or missing: <OID_PATH>
    [ the rest of the fsck output here, i.e. it didn't hard die ]

We'll of course still exit with non-zero, but now we'll finish the
rest of the traversal. The tests that's being added here asserts that
we'll still complain about other fsck issues (e.g. an unrelated
dangling blob).

But why are we complaining about a "hash mismatch" for an object of a
type we don't know about? We shouldn't, and the next commit will fix
that. This is the bare minimal change to not make fsck hard die on the
issue, in the next commit we'll teach fsck to recognize this
particular type of corruption and emit a better error message.

The parse_loose_header() function being changed here is only used in
builtin/fsck.c, see f6371f92104 (sha1_file: add read_loose_object()
function, 2017-01-13) for its introduction.

Perhaps we'll end up moving it there in the future, but having it in
the utility belt seems like a good thing even if nothing else is using
it. But let's at least move its changed prototype from cache.h while
we're at it.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>