fsck tests: refactor one test to use a sub-repo
Refactor one of the fsck tests to use a sub-repo. It's a pervasive
pattern in t1450-fsck.sh to spend a lot of effort on the teardown of a
tests so we're not leaving corrupt content for the next test.

We should instead simply use something like this test_create_repo
pattern. It's both less verbose, and makes things easier to debug as a
failing test will have its corrupt content left under -d without
damaging the state for other tests.

But I'm only going to add another version of this test in a follow-up
commit, so skip that general refactoring of the file, and just change
this one test to use the saner pattern.

I'm also not changing the "sha" variable to "oid" just so that this
diff is more readable under "-w".

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>